Rename publish application service package
